// src/lib/api/models/createNotificationTemplateDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CreateNotificationTemplateDtoChannel } from './createNotificationTemplateDtoChannel';

export interface CreateNotificationTemplateDto {
  code: string;
  titleTemplate: string;
  bodyTemplate: string;
  channel?: CreateNotificationTemplateDtoChannel;
}