// src/lib/api/models/updateNotificationTemplateDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { UpdateNotificationTemplateDtoChannel } from './updateNotificationTemplateDtoChannel';

export interface UpdateNotificationTemplateDto {
  code?: string;
  titleTemplate?: string;
  bodyTemplate?: string;
  channel?: UpdateNotificationTemplateDtoChannel;
}