// src/lib/api/models/createNotificationTemplateDtoChannel.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type CreateNotificationTemplateDtoChannel = typeof CreateNotificationTemplateDtoChannel[keyof typeof CreateNotificationTemplateDtoChannel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateNotificationTemplateDtoChannel = {
  EMAIL: 'EMAIL',
  SMS: 'SMS',
  PUSH: 'PUSH',
  IN_APP: 'IN_APP',
} as const;